+++ client/src/Template.js
@@ -2,6 +2,7 @@ import styled from 'styled-components';
 
 const StyledTemplate = styled.div`
   width: 100%;
+  font-size: 2rem;
   @media screen and (min-width: 600px) {
     & > * {
       margin: 5% 10% !important;
